fix: fix social auth tests and SQLite compatibility for SocialAccount model

+++ backend/auth-service/app/models/social_account.py
@@ -1,5 +1,5 @@
 import os
-from sqlalchemy import Column, String, DateTime, BigInteger, UniqueConstraint
+from sqlalchemy import Column, String, DateTime, BigInteger, Integer, UniqueConstraint
 from datetime import datetime
 from app.db.base import Base
 
@@ -13,8 +13,8 @@ class SocialAccount(Base):
         {"schema": DB_SCHEMA} if DB_SCHEMA else {},
     )
 
-    id = Column(BigInteger, primary_key=True, index=True)
-    user_id = Column(BigInteger, nullable=False)
+    id = Column(Integer, primary_key=True, index=True)
+    user_id = Column(Integer, nullable=False)
     provider = Column(String(20), nullable=False)
     provider_user_id = Column(String(255), nullable=False)
     created_at = Column(DateTime(timezone=True), default=datetime.utcnow, nullable=False)
